La partie read du crud est terminée pour operator et patient.

+++ app/src/main/java/com/example/ex_motricite/PatientDAO.java
@@ -13,7 +13,7 @@ public class PatientDAO {
         accesBD = new BdSQLiteOpenHelper(ct, base, null, version);
     }
 
-    public Patient getPatient(int idPatient){
+    public Patient getPatient(long idPatient){
         Patient patient = null;
         Cursor curseur;
         curseur = accesBD.getReadableDatabase().rawQuery("select * from patient where idPatient="+idPatient+";",null);
